refactor(heap): relocate demo functions to their respective implementation files
